Migrate file upload to new fileUpload LWC
